Matcher 2026 07 30 15 30 00

+++ include/functions.us.h
@@ -759,7 +759,7 @@ void func_800E72A0_F6250(void);
 void func_800E7AA0_F6A50(VehicleInstance *arg0);
 void func_800E8890_F7840(void);
 s32 func_800703B0_7F360(s16 arg0, s16 arg1);
-s32 func_800B9228_C81D8(s32 arg0, s16 arg1, s32 arg2, s16 arg3, s16 arg4);
+s32 func_800B9228_C81D8(s16 arg0, s16 arg1, s16 arg2, s16 arg3, s16 arg4);
 s32 func_800B93AC_C835C(s16 arg0, s16 arg1, s32 arg2, s16 arg3, s32 arg4, s32 arg5);
 void func_800DC18C_EB13C(Vec3f *arg0, u8 *arg1, u8 *arg2, u16 arg3, u8 arg4);
 void func_800DC5B8_EB568(Vec3f *arg0, f32 arg1, s32 arg2, s32 arg3);
